final commit

+++ API/Infrastructure/Services/BookService.cs
@@ -6,6 +6,7 @@ using Infrastructure.Data;
 using Infrastructure.Exceptions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
+using System.Linq;
 
 namespace Infrastructure.Services
 {
@@ -49,6 +50,113 @@ namespace Infrastructure.Services
                 throw new Exception("Có lỗi xảy ra khi lấy thông tin sách", ex);
             }
         }
+        public async Task<IEnumerable<BookDTO>> GetNewInfoBooksAsync(string arrBookID)
+        {
+            try
+            {              
+                var bookIDs = arrBookID.Split(", ");
+                var newInfoBooks = await _context.Books
+                    .Where(b => bookIDs.Contains(b.BookID.ToString()))
+                    .ToListAsync();
+
+                return _mapper.Map<IEnumerable<BookDTO>>(newInfoBooks);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Có lỗi khi lấy sách liên quan", ex);
+            }
+        }
+
+        public async Task<object> GetTopBooksAsync()
+        {
+            try
+            {
+                var newestBooks = await _context.Books
+                    .OrderByDescending(b => b.PublishedDate)
+                    .Take(4)
+                    .ToListAsync();
+
+                var bestSellingBooks = await _context.OrderDetails
+                    .Where(od => od.Order.OrderStatus != "Canceled")
+                    .GroupBy(od => od.BookID)
+                    .Select(g => new
+                    {
+                        BookID = g.Key,
+                        BookName = g.First().Book.BookName,
+                        Author = g.First().Book.Author,
+                        Publisher = g.First().Book.Publisher,
+                        PublishedDate = g.First().Book.PublishedDate,
+                        Price = g.First().Book.Price,
+                        StockQuantity = g.First().Book.StockQuantity,
+                        Description = g.First().Book.Description,
+                        ImagePath = g.First().Book.ImagePath,
+                        Categories = g.First().Book.Categories,
+                        TotalQuantitySold = g.Sum(od => od.Quantity)
+                    })
+                    .OrderByDescending(b => b.TotalQuantitySold)
+                    .Take(4)
+                    .ToListAsync();       
+                return new
+                {
+                    TopNew = newestBooks,
+                    TopBestSell = bestSellingBooks
+                };
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Có lỗi xảy ra khi lấy danh sách top sách", ex);
+            }
+        }
+        public async Task<IEnumerable<BookDTO>> SearchBooksAsync(string searchName)
+        {
+            try
+            {
+                var resultBooks = await _context.Books
+                                         .Where(b => b.BookName.Contains(searchName))
+                                         .AsNoTracking()
+                                         .ToListAsync();
+                return _mapper.Map<IEnumerable<BookDTO>>(resultBooks);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Có lỗi khi lấy sách liên quan", ex);
+            }
+        }
+
+
+        public async Task<IEnumerable<BookDTO>> GetRelatedBooksAsync(int bookID)
+        {
+            try
+            {
+                var currentBook = await _context.Books
+                    .Where(b => b.BookID == bookID)
+                    .FirstOrDefaultAsync();
+
+                if (currentBook == null)
+                {
+                    throw new NotFoundException("Sách không tồn tại");
+                }
+
+                var categories = currentBook.Categories.Split(", ");
+
+                var relatedBooks = await _context.Books
+                    .Where(b => b.BookID != bookID) 
+                    .ToListAsync(); 
+
+                var relatedBooksFiltered = relatedBooks
+                    .Where(b => b.Categories.Split(", ").Intersect(categories).Any())
+                    .OrderByDescending(b => b.PublishedDate)
+                    .Take(4)
+                    .ToList();
+
+                return _mapper.Map<IEnumerable<BookDTO>>(relatedBooks);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Có lỗi khi lấy sách liên quan", ex);
+            }
+        }
+
 
         public async Task<BookDTO> CreateBookAsync(BookDTO bookDto)
         {
@@ -85,7 +193,7 @@ namespace Infrastructure.Services
                 var book = await _context.Books.FindAsync(id)
                     ?? throw new NotFoundException("Không tìm thấy sách");
 
-                Validate(bookDto);              
+                Validate(bookDto);
 
                 book.BookName = bookDto.BookName;
                 book.Author = bookDto.Author;
@@ -122,12 +230,27 @@ namespace Infrastructure.Services
                 var book = await _context.Books.FindAsync(id)
                     ?? throw new NotFoundException("Không tìm thấy sách");
 
+                var isUsedInCart = await _context.ShoppingCarts
+            .           AnyAsync(cart => cart.BookID == id);
+
+                var isUsedInOrders = await _context.OrderDetails
+                    .AnyAsync(od => od.BookID == id);
+
+                if (isUsedInCart || isUsedInOrders)
+                {
+                    throw new BadRequestException("Không thể xóa sách vì nó đang được sử dụng trong giỏ hàng hoặc đơn hàng."); ;
+                }
                 _context.Books.Remove(book);
                 await _context.SaveChangesAsync();
 
                 return new { message = "Xóa sách thành công!" };
             }
             catch (NotFoundException)
+            {
+                throw;
+            }
+
+            catch (BadRequestException)
             {
                 throw;
             }
@@ -162,11 +285,11 @@ namespace Infrastructure.Services
 
                 throw new Exception("Có lỗi xảy ra khi thêm hình ảnh " + ex.Message, ex);
             }
-            
+
         }
 
         private static void Validate(BookDTO bookDto)
-        {          
+        {
 
             if (bookDto.Price < 0)
             {
